Moved cache tests into syntax_tests/test_cache.py.
